test: put validation functions in separate files
